Major: Improving the code coverage + creating search component
