+++ src/app/scheduling/create/scheduling-create.component.spec.ts
@@ -44,56 +44,6 @@ describe('SchedulingCreateComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('cityFilter should apply a regex on untreated string', () => {
-    const city = {
-      name: <string> 'BRASILIA',
-      code: <string> '5300108'
-    };
-    const dataFromAPI = '"5300108:BRASILIA"';
-    expect(component.cityFilter(dataFromAPI)).toEqual(city);
-  });
-
-  it('cityPush should push a filtered city into a array', () => {
-    var dataFromAPI:string[];
-    dataFromAPI = ['5300108:BRASILIA'];
-
-
-    const city = {
-      name: 'BRASILIA',
-      code: <string> '5300108'
-    }
-
-    var cities = Array<Object>();
-    cities.push(city);
-    expect(component.cityPush(dataFromAPI)).toEqual(cities);
-  });
-
-  it('filterSchools should return a Array<Object>', () => {
-    const school = {
-      anoCenso: '2013',
-      cidade: 'BRASILIA',
-      cod: '53012097',
-      codCidade: '5300108',
-      dependenciaAdministrativa: '2',
-      dependenciaAdministrativaTxt: 'Estadual',
-      enemMediaGeral: '472.6940002441406',
-      estado: 'DF',
-      idebAF: '2.9000000953674316',
-      idebAI: '0',
-      nome:  'CED 123 DE SAMAMBAIA',
-      regiao: 'Centro-Oeste',
-      situacaoFuncionamento: '1',
-      situacaoFuncionamentoTxt: 'Em atividade'
-    };
-
-    var schools = [];
-    var afterTreatment = [];
-    afterTreatment.push(school);
-    schools.push(1);
-    schools.push(school);
-    expect(component.filterSchools(schools)).toEqual(afterTreatment[0]);
-
-  });
   it('should navigate to "agendamento"', fakeAsync(() => {
     location = TestBed.get(Location);
     component.navigate();
@@ -116,9 +66,4 @@ describe('SchedulingCreateComponent', () => {
     expect(location.path()).toBe('/agendamento');
 
   });
-  it('should clear on click', () => {
-    component.searchSchool();
-    expect(component.search.situation).toEqual('1');
-
-      });
 });
