+++ src/app/scheduling/create/scheduling-create.component.ts
@@ -23,7 +23,6 @@ export class SchedulingCreateComponent implements OnInit {
 
   constructor(
     private schedulingService: SchedulingService,
-    private schoolService: SchoolService,
     private router: Router,
     private alertService: AlertService
   ) {}
@@ -38,87 +37,14 @@ export class SchedulingCreateComponent implements OnInit {
     this.search = new Search();
   }
 
-  searchSchool(): void {
-    this.search.state = this.state;
-    this.search.situation = '1';
-
-    this.schoolService.searchSchool(this.search)
-      .subscribe(
-          result => {
-            this.schools = this.filterSchools(result);
-          },
-          error => {
-            alert(error);
-            console.error(error);
-      });
-  }
-
-  filterSchools(result: Array<Object>): Array<Object> {
-    const res = [];
-
-    result.forEach(subitem => {
-      res.push(subitem);
-    });
-
-    console.log('Resultado da busca: ' , res[1]);
-    return res[1];
-  }
-
-  searchCity(): void {
-    this.schoolService.searchCity(this.state)
-      .subscribe(
-          result => {
-            this.cities = this.cityPush(result);
-            console.log(this.cities);
-          },
-          error => {
-            alert(error);
-            console.error(error);
-      });
-  }
-
-  cityPush(result: Array<Object>): Array<Object> {
-    console.log(result);
-
-    const cities = [];
-
-    result.forEach(subitem => {
-      const untreated = JSON.stringify(subitem);
-      const city = this.cityFilter(untreated);
-      cities.push(city);
-    });
-
-    return cities;
-  }
-
-  cityFilter(untreated: string): Object {
-    const city = {
-      name: <string> null,
-      code: <string> null
-    };
-    city.name = untreated;
-    console.log(city.name);
-
-    const quote = /\"/g;
-    const letters = /[\d:-]+/g;
-    const numbers = /:[A-Za-záàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]+$/;
-    const removeQuote = city.name.replace(quote, '');
-
-    city.name = removeQuote.replace(letters, '');
-    city.code = removeQuote.replace(numbers, '');
-    console.log(city);
-
-    return city;
-  }
+ 
   newScheduling(): void {
     if (this.formScheduling.form.valid) {
       this.schedulingService.newScheduling(this.scheduling);
       this.router.navigate(['/agendamento']);
     }
   }
-  toggleCollapsed(): void {
-    this.collapsed = !this.collapsed;
-  }
+
   navigate() {
     this.router.navigate(['/agendamento']);
   }
