+++ src/app/scheduling/search/search.component.html
@@ -1,0 +1,73 @@
+<br><br><br><br>
+
+<p class="h2 text-center mb-6"> Criar Agendamento </p>
+
+<br>
+
+<div class="create-form">
+    <div>
+        <div class="card container search">
+            <div class="d-flex justify-content-end">
+                <p class="mr-auto p-2 h6 text-center">Buscar Escolas</p>
+                <button class="btn btn-deep-orange p-2" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation" (click)="toggleCollapsed()">
+          <span class="fa fa-chevron-down" ></span>
+        </button>
+            </div>
+
+            <div [ngClass]="{'collapse': collapsed, 'navbar-collapse': true}">
+
+                <label for="name"><i class="fa fa-globe" aria-hidden="true"></i> Sigla do Estado:</label>
+                <div class="input-group">
+                    <input type="text" class="form-control" id="state" name="state" [(ngModel)]="state" #stt="ngModel" minlength="2" maxlength="2">
+                    <span class="input-group-btn">
+              <button class="btn btn-deep-orange rounded" type="button" aria-label="Search"
+              (click)="searchCity()">
+              <i class="fa fa-search" aria-hidden="true"></i>
+            </button>
+          </span>
+                </div>
+
+                <div class="school-form">
+                    <label><i class="fa fa-road" aria-hidden="true"></i> Municipio:</label>
+                    <div class="container select-style">
+                        <select [(ngModel)]="search.city">
+            <option value="" selected disabled>Escolha um Município</option>
+            <option *ngFor="let c of cities" [value]="c.code">{{c.name}}</option>
+          </select>
+                    </div>
+                    <label for="name"><i class="fa fa-info" aria-hidden="true"></i> Nome da Escola:</label>
+                    <input type="text" class="form-control" id="name" name="name" [(ngModel)]="search.name" #name="ngModel">
+
+                    <label for="code"><i class="fa fa-pencil" aria-hidden="true"></i> Código da Escola:</label>
+                    <input type="text" class="form-control" id="code" name="code" [(ngModel)]="search.code" #code="ngModel">
+                </div>
+                <div class="text-center">
+                    <input type="button" class="btn btn-deep-orange" (click)="searchSchool()" value="Buscar Escolas">
+                </div>
+                <table *ngIf="schools.length>0" class="table">
+                    <thead class="deep-orange lighten-4">
+                        <tr>
+                            <th>#</th>
+                            <th>Nome</th>
+                            <th>Código</th>
+                            <th>Rede</th>
+                            <th>Cidade</th>
+                            <th>Estado</th>
+                        </tr>
+                    </thead>
+
+                    <tbody>
+                        <tr *ngFor="let school of schools; let i = index ">
+                            <th scope="row">{{i + 1}}</th>
+                            <td>{{school.nome}}</td>
+                            <td>{{school.cod}}</td>
+                            <td>{{school.dependenciaAdministrativaTxt}}</td>
+                            <td>{{school.cidade}}</td>
+                            <td>{{school.estado}}</td>
+                        </tr>
+                    </tbody>
+                </table>
+            </div>
+        </div> 
+    </div>
+</div>
